+++ lib/Service/ImapLoader.php
@@ -21,7 +21,12 @@
 namespace OCA\Majordomo\Service;
 
 use DateTime;
+use DirectoryTree\ImapEngine\Address;
+use DirectoryTree\ImapEngine\FolderInterface;
+use DirectoryTree\ImapEngine\Mailbox;
+use DirectoryTree\ImapEngine\MessageInterface;
 use OCA\Majordomo\Db\MailingList;
+use OCP\IDateTimeFormatter;
 use Psr\Log\LoggerInterface;
 
 class ImapLoader {
@@ -30,7 +35,7 @@ class ImapLoader {
     const BOUNCE_PATTERN = "/BOUNCE +([^@]*@[^:]*): +Non-member submission from \[([^]]*)]/";
 
     private $AppName;
-    private $imap = [];
+    private Mailbox|NULL $imap = NULL;
     private $imapSettings;
     /**
      * @var DateTime
@@ -48,12 +53,14 @@ class ImapLoader {
      * @var ResendService
      */
     private $resendService;
+    private IDateTimeFormatter $dateTimeFormatter;
 
     function __construct(
         $AppName,
         Settings $settings,
         InboundService $inboundService,
         ResendService $resendService,
+        IDateTimeFormatter $dateTimeFormatter,
         LoggerInterface $logger
     ) {
         $this->imapSettings = $settings->getImapSettings();
@@ -61,67 +68,78 @@ class ImapLoader {
         $this->inboundService = $inboundService;
         $this->AppName = $AppName;
         $this->resendService = $resendService;
+        $this->dateTimeFormatter = $dateTimeFormatter;
     }
 
     public function isEnabled() {
         return $this->imapSettings !== NULL && !empty($this->imapSettings->server);
     }
 
-    public function getImap($folder = NULL) {
+    public function getImap($folder = NULL): Mailbox {
         if ($folder == NULL) {
             $folder = $this->imapSettings->inbox;
         }
 
-        if (!isset($this->imap[$folder])) {
+        if ($this->imap == NULL) {
             $mailbox = '{' . $this->imapSettings->server . '}' . $folder;
             $this->logger->debug("Opening IMAP connection: $mailbox");
-            $imap = imap_open(
-                $mailbox, $this->imapSettings->user, $this->imapSettings->password
-            );
-            if ($imap !== false) {
-                $this->imap[$folder] = $imap;
-            } else {
-                $this->logger->error("Failed to open IMAP connection to $mailbox: " . imap_last_error());
-                return false;
+            $serverParts = explode("/", strtolower($this->imapSettings->server));
+            $hostParts = explode(":", array_shift($serverParts));
+            $config = [
+                "host" => $hostParts[0],
+                "username" => $this->imapSettings->user,
+                "password" => $this->imapSettings->password,
+            ];
+
+            if (isset($hostParts[1])) {
+                $config["port"] = $hostParts[1];
             }
+
+            if (in_array("tls", $serverParts)) {
+                $config["encryption"] = "starttls";
+            } elseif (in_array("ssl", $serverParts)) {
+                $config["encryption"] = "ssl";
+            } else {
+                $config["encryption"] = null;
+            }
+
+            if (in_array("novalidate-cert", $serverParts)) {
+                $config["validate_cert"] = false;
+            }
+
+            $this->imap = new Mailbox($config);
+            $this->imap->connect();
         }
 
-        return $this->imap[$folder];
+        return $this->imap;
     }
 
     function __destruct() {
-        foreach (array_values($this->imap) as $imap) {
-            imap_close($imap);
+        if ($this->imap != NULL) {
+            $this->imap->disconnect();
         }
-        $this->imap = [];
+        $this->imap = NULL;
     }
 
     public function test() {
-        $imap = $this->getImap();
-        if (!$imap) {
-            throw new \RuntimeException("IMAP connection failed: " . imap_last_error());
-        }
-
         return [
             "folders" => $this->getFolders()
         ];
     }
 
-    private function getFolders() {
+    private function getFolders(): array {
         $imap = $this->getImap();
-        $ref = '{' . $this->imapSettings->server . '}';
-        return array_map(function ($folder) use ($ref) {
-            return strpos($folder, $ref) === 0 ? substr($folder, strlen($ref)) : $folder;
-        }, imap_list($imap, $ref, '*'));
+        return $imap->folders()->get()->map(function (FolderInterface $folder): string {
+            return $folder->name();
+        })->toArray();
     }
 
     private function ensureFoldersExist() {
         $imap = $this->getImap();
-        $ref = '{' . $this->imapSettings->server . '}';
         $folders = $this->getFolders();
         foreach ([ $this->imapSettings->archive, $this->imapSettings->errors, $this->imapSettings->bounces ] as $required) {
             if ($required !== NULL && !in_array($required, $folders)) {
-                imap_createmailbox($imap, $ref . $required);
+                $imap->folders()->create($required);
             }
         }
     }
@@ -143,142 +161,165 @@ class ImapLoader {
 
         return $results;
     }
-    
-    private function fetchMails() {
+
+    /**
+     * @return MessageInterface[]
+     */
+    private function fetchMails(): array {
         $imap = $this->getImap();
-        $out = array();
         $this->ensureFoldersExist();
-        $nrs = imap_sort($imap, SORTARRIVAL, 1);
-        if ($nrs !== false) {
-            $mails = array();
-            foreach (imap_fetch_overview($imap, join(",", $nrs)) as $mail) {
-                $mails[$mail->msgno] = $mail;
-            }
-            foreach ($nrs as $nr) {
-                $out[] = $mails[$nr];
-            }
-        }
-        return $out;
+        $folder = $imap->folders()->firstOrCreate($this->imapSettings->inbox);
+        return $folder->messages()->all()->setFetchOrderDesc()->withBody()->withHeaders()->get()->all();
     }
     
     public function processMails() {
         $imap = $this->getImap();
         $expunge = false;
         foreach ($this->fetchMails() as $mail) {
-            $from = strtolower($mail->from);
-            $this->logger->debug("Checking inbound mail from {$mail->from} with subject '{$mail->subject}'");
-            if (strncasecmp($mail->subject, self::SUBJECT_PREFIX, strlen(self::SUBJECT_PREFIX)) == 0) {
-                try {
-                    $requestId = substr($mail->subject, strlen(self::SUBJECT_PREFIX));
-                    $results = $this->parseMailBody(imap_body($imap, $mail->msgno));
-                    $this->inboundService->handleResult($requestId, $results, $from);
-                    imap_mail_move($imap, $mail->msgno, $this->imapSettings->archive);
-                    $this->logger->info("Processed mail {$mail->msgno} '{$mail->subject}' from {$mail->from}", [ "app" => $this->AppName ]);
-                } catch (\Exception $e) {
-                    $this->logger->error("Failed to process mail {$mail->msgno} '{$mail->subject}' from {$mail->from}", [
-                        "app" => $this->AppName,
-                        "exception" => $e
-                    ]);
-                    imap_mail_move($imap, $mail->msgno, $this->imapSettings->errors);
-                }
-                $expunge = true;
-            } else if (!empty($this->imapSettings->bounces) && preg_match(self::BOUNCE_PATTERN, $mail->subject)) {
-                imap_mail_move($imap, $mail->msgno, $this->imapSettings->bounces);
-                $this->logger->info("Moved bounce {$mail->msgno} '{$mail->subject}' from {$mail->from}", [ "app" => $this->AppName ]);
-                $expunge = true;
-            } else if ($this->resendService->isEnabled()) {
-                $header = imap_headerinfo($imap, $mail->msgno);
-                if (isset($header->to)) {
-                    $this->logger->debug("header: " . print_r($header, true));
-                    $to = array_map(function ($t) {
-                        return strtolower($t->mailbox . '@' . $t->host);
-                    }, $header->to);
-
-                    $rawHeader = null;
-                    $body = null;
-                    foreach ($this->resendService->getLists($to) as $ml) {
-                        if ($body === null) {
-                            $rawHeader = imap_fetchheader($imap, $mail->msgno);
-                            $body = imap_body($imap, $mail->msgno);
-                        }
-                        $this->logger->info("Resending mail from $from to " . $ml->title, ["app" => $this->AppName]);
-                        $this->resendService->bounceOrResend($ml, $from, $rawHeader, $body);
-                    }
-                    if ($body !== null) {
-                        imap_mail_move($imap, $mail->msgno, $this->imapSettings->archive);
-                        $expunge = true;
-                    }
-                }
-            }
+            $expunge = $this->processMail($mail) || $expunge;
         }
 
         if ($expunge) {
-            imap_expunge($imap);
+            $imap->folders()->firstOrCreate($this->imapSettings->inbox)->expunge();
         }
     }
 
+    public function idle() {
+        $imap = $this->getImap();
+        $this->logger->info("Waiting for new messages");
+        $inbox = $imap->folders()->firstOrCreate($this->imapSettings->inbox);
+        $inbox->idle(function (MessageInterface $message) use ($inbox) {
+            $message = $inbox->messages()->uid($message->uid())->withBody()->withHeaders()->withFlags()->first();
+            if ($message !== null) {
+                $this->logger->info("Got new message: {$message->subject()}");
+                $this->processMail($message);
+            }
+        });
+    }
+
     public function getBounces() {
-        $imap = $this->getImap($this->imapSettings->bounces);
+        $imap = $this->getImap();
         $out = array();
         $this->ensureFoldersExist();
-        $nrs = imap_sort($imap, SORTARRIVAL, 1);
+        /** @var MessageInterface[] $bounceMails */
+        $bounceMails = $imap->folders()->firstOrCreate($this->imapSettings->bounces)
+            ->messages()->all()->setFetchOrderDesc()
+            ->withBody()->withHeaders()->withFlags()
+            ->get()->all();
         $bouncerMapping = $this->inboundService->getBouncerMapping();
-        if ($nrs !== false) {
-            $mails = [];
-            foreach (imap_fetch_overview($imap, join(",", $nrs)) as $mail) {
-                $m = [];
-                if (isset($mail->subject) && preg_match(self::BOUNCE_PATTERN, $mail->subject, $m)) {
-                    if (!array_key_exists($m[1], $bouncerMapping)) {
-                        continue;
-                    }
-
-                    if ($mail->deleted) {
-                        continue;
-                    }
-
-                    $ml = $bouncerMapping[$m[1]];
-                    $mails[$mail->msgno] = [
-                        "list_id" => $ml->id,
-                        "list_title" => $ml->title,
-                        "list_address" => $m[1],
-                        "from" => $m[2],
-                        "date" => $mail->date,
-                        "mid" => $mail->message_id,
-                        "uid" => $mail->uid,
-                    ];
+        foreach ($bounceMails as $mail) {
+            $m = [];
+            if ($mail->subject() != null && preg_match(self::BOUNCE_PATTERN, $mail->subject(), $m)) {
+                if (!array_key_exists($m[1], $bouncerMapping)) {
+                    continue;
                 }
-            }
-            foreach ($nrs as $nr) {
-                if (isset($mails[$nr])) {
-                    $out[] = $mails[$nr];
+
+                if ($mail->hasFlag("Deleted")) {
+                    continue;
                 }
+
+                $ml = $bouncerMapping[$m[1]];
+                $out[] = [
+                    "list_id" => $ml->id,
+                    "list_title" => $ml->title,
+                    "list_address" => $m[1],
+                    "from" => $m[2],
+                    "date" => $this->dateTimeFormatter->formatDateTime($mail->date()->toDate()),
+                    "mid" => $mail->messageId(),
+                    "uid" => $mail->uid(),
+                ];
             }
         }
         return $out;
     }
 
     public function getBounce($uid) {
-        $imap = $this->getImap($this->imapSettings->bounces);
+        $imap = $this->getImap();
         $ml = $this->assertBounce($uid);
+        $this->logger->info("Bounce: " . $uid);
+        $body = (string)$imap->folders()->firstOrCreate($this->imapSettings->bounces)
+            ->messages()->uid($uid)
+            ->withHeaders()->withBody()
+            ->firstOrFail();
+        $this->logger->info("Mail: ". $body);
         return [
             "ml" => $ml,
-            "body" => imap_body($imap, $uid, FT_UID),
+            "body" => $body,
         ];
     }
 
     public function deleteBounce($uid) {
-        $imap = $this->getImap($this->imapSettings->bounces);
+        $imap = $this->getImap();
         $this->assertBounce($uid);
-        imap_mail_move($imap, $uid, $this->imapSettings->archive, FT_UID);
+        $imap->folders()->firstOrCreate($this->imapSettings->bounces)
+            ->messages()
+            ->uid($uid)
+            ->firstOrFail()
+            ->move($this->imapSettings->archive);
     }
 
     protected function assertBounce($uid): MailingList {
-        $imap = $this->getImap($this->imapSettings->bounces);
-        $mail = imap_fetch_overview($imap, "$uid", FT_UID)[0];
+        $imap = $this->getImap();
+        $mail = $imap->folders()->firstOrCreate($this->imapSettings->bounces)
+            ->messages()
+            ->uid($uid)
+            ->withHeaders()
+            ->firstOrFail();
         $m = [];
-        if (preg_match(self::BOUNCE_PATTERN, $mail->subject, $m)) {
+        if (preg_match(self::BOUNCE_PATTERN, $mail->subject(), $m)) {
             return $this->inboundService->getListByBounceAddress($m[1]);
         }
         throw new \RuntimeException("The mail $uid is not a bounced message");
+    }
+
+    protected function processMail(MessageInterface $mail): bool {
+        try {
+            $expunge = false;
+            $from = strtolower($mail->from()->email());
+            $this->logger->debug("Checking inbound mail from {$mail->from()->email()} with subject '{$mail->subject()}'");
+            if (strncasecmp($mail->subject(), self::SUBJECT_PREFIX, strlen(self::SUBJECT_PREFIX)) == 0) {
+                $requestId = substr($mail->subject(), strlen(self::SUBJECT_PREFIX));
+                $results = $this->parseMailBody($mail->text());
+                $this->inboundService->handleResult($requestId, $results, $from);
+                $mail->move($this->imapSettings->archive);
+                $this->logger->info("Processed mail {$mail->messageId()} '{$mail->subject()}' from {$mail->from()->email()}", ["app" => $this->AppName]);
+                $expunge = true;
+            } else if (!empty($this->imapSettings->bounces) && preg_match(self::BOUNCE_PATTERN, $mail->subject())) {
+                $mail->move($this->imapSettings->bounces);
+                $this->logger->info("Moved bounce {$mail->messageId()} '{$mail->subject()}' from {$mail->from()->email()}", ["app" => $this->AppName]);
+                $expunge = true;
+            } else if ($this->resendService->isEnabled()) {
+                $to = array_map(function (Address $address) {
+                    return $address->email();
+                }, $mail->to());
+                if (!empty($to)) {
+                    $resent = false;
+                    foreach ($this->resendService->getLists($to) as $ml) {
+                        $this->logger->info("Resending mail from $from to " . $ml->title, ["app" => $this->AppName]);
+                        $this->resendService->bounceOrResend($ml, $mail);
+                        $resent = true;
+                    }
+                    if ($resent) {
+                        $mail->move($this->imapSettings->archive);
+                        $expunge = true;
+                    }
+                }
+            }
+            return $expunge;
+        } catch (\Exception $e) {
+            try {
+                $this->logger->error("Failed to process mail {$mail->messageId()} '{$mail->subject()}' from {$mail->from()->email()}", [
+                    "app" => $this->AppName,
+                    "exception" => $e
+                ]);
+            } catch (\Exception $ignored) {
+                $this->logger->error("Failed to process mail $mail", [
+                    "app" => $this->AppName,
+                    "exception" => $e
+                ]);
+            }
+            $mail->move($this->imapSettings->errors);
+            return true;
+        }
     }
 }